+++ src/DependencyInjection/ThreeBRSSyliusSortingExtension.php
@@ -4,21 +4,22 @@ declare(strict_types=1);
 
 namespace ThreeBRS\SortingPlugin\DependencyInjection;
 
+use Sylius\Bundle\ResourceBundle\DependencyInjection\Extension\AbstractResourceExtension;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 /**
  * This is the class that loads and manages your bundle configuration.
  *
  * @see http://symfony.com/doc/current/cookbook/bundles/extension.html
  */
-class ThreeBRSSyliusSortingExtension extends Extension
+class ThreeBRSSyliusSortingExtension extends AbstractResourceExtension
 {
+    /** @psalm-suppress UnusedVariable */
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
     }
 }